+++ streamer/CHANGELOG.md
@@ -11,6 +11,11 @@ All notable changes to this project will be documented in this file.
 * Support for Swift Package Manager (contributed by [@stevenzeck](https://github.com/readium/r2-streamer-swift/pull/210)).
 * EPUB publications implement a `SearchService` to search through the content.
 
+### Changed
+
+* The default EPUB positions service now uses the archive entry length when available. [This is similar to how Adobe RMSDK generates page numbers](https://github.com/readium/architecture/issues/123).
+    * To use the former strategy, create the `Streamer` with: `Streamer(parsers: [EPUBParser(reflowablePositionsStrategy: .originalLength(pageLength: 1024))])`
+    
 
 ## [2.0.0]